+++ handlers/handler.go
@@ -1,0 +1,44 @@
+package handlers
+
+import (
+	"TODO_GO/models"
+	"encoding/json"
+	"net/http"
+)
+
+func TodosHandler(w http.ResponseWriter, r *http.Request) {
+	switch r.Method {
+	case http.MethodGet:
+		getTodos(w)
+	case http.MethodPost:
+		createTodo(w, r)
+	default:
+		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
+	}
+}
+
+// GET /todos
+func getTodos(w http.ResponseWriter) {
+	todos, err := models.GetTodos()
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+	w.Header().Set("Content-Type", "application/json")
+	json.NewEncoder(w).Encode(todos)
+}
+
+// POST /todos
+func createTodo(w http.ResponseWriter, r *http.Request) {
+	var t models.Todo
+	if err := json.NewDecoder(r.Body).Decode(&t); err != nil {
+		http.Error(w, err.Error(), http.StatusBadRequest)
+		return
+	}
+	if err := models.CreateTodo(&t); err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+	w.Header().Set("Content-Type", "application/json")
+	json.NewEncoder(w).Encode(t)
+}
